# swagger.yml
basePath: /
consumes:
- application/json
- application/xml
host: 127.0.0.1:6420
info:
  contact:
    email: skycoin.doe@example.com
    name: Skycoin project
    url: http://127.0.0.1:6420
  description: Skycoin is a next-generation cryptocurrency.
  title: Skycoin REST API.
  version: 0.25.1
paths:
...
  /api/v2/address/verify:
    post:
      operationId: verifyAddress
      parameters:
      - description: Address id.
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response verifies a Skycoin address
          schema:
            properties:
              data:
                properties:
                  version:
                    format: int64
                    type: integer
                type: object
              error:
                type: object
            type: object
        default:
          $ref: '#/responses/genericError'
      security:
      - csrfAuth: []
      summary: Verifies a Skycoin address.
...
produces:
- application/json
- application/xml
responses:
...
  genericError:
    description: |-
      A GenericError is the default error message that is generated.
      For certain status codes there are more appropriate error structures.
    schema:
      properties:
        code:
          format: int32
          type: integer
          x-go-name: Code
        message:
          type: string
          x-go-name: Message
      type: object
schemes:
- http
- https
securityDefinitions:
  csrfAuth:
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"